requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
llama-cloud
httpx
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
streamlit
streamlit-option-menu
pandas
requests
pinecone
sentence-transformers
pytest

